% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_play_by_play_modifiers.R
\name{add_date_columns}
\alias{add_date_columns}
\title{add the date information to the game play by play table}
\usage{
add_date_columns(game_play_by_play, game_id)
}
\arguments{
\item{game_play_by_play}{the game play by play table}

\item{game_id}{the id of the game from which to draw date information}
}
\value{
the game play by play table with date and time info added
}
\description{
add the date information to the game play by play table
}
\keyword{internal}
