% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{add_nflscrapR_variables}
\alias{add_nflscrapR_variables}
\title{calculate ep and wp using the functions from nflscrapR}
\usage{
add_nflscrapR_variables(game_pbp)
}
\arguments{
\item{game_pbp}{play by play table}
}
\value{
play by play table with columns for ep and wp stats
}
\description{
calculate ep and wp using the functions from nflscrapR
}
\keyword{internal}
