% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{convert_types_for_nflscrapR}
\alias{convert_types_for_nflscrapR}
\title{Converts integer columns to doubles so that nflscrapR ep and wp functions work properly}
\usage{
convert_types_for_nflscrapR(game_pbp)
}
\arguments{
\item{game_pbp}{play by play table}
}
\value{
the play by play table with the columns having the correct class for the ep and wp formulas
}
\description{
Converts integer columns to doubles so that nflscrapR ep and wp functions work properly
}
\keyword{internal}
