% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_data.R
\name{create_binary_indicators}
\alias{create_binary_indicators}
\title{fill in indicator variable in the play by play table for a given play}
\usage{
create_binary_indicators(play_number, game_environment)
}
\arguments{
\item{play_number}{the index of the play in the play_stat_matrix_list}

\item{game_environment}{contains preprocessed structures for analyzing the game}
}
\value{
a row of the play by play table with indicators filled in
}
\description{
fill in indicator variable in the play by play table for a given play
}
\keyword{internal}
