% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{download_and_parse_game}
\alias{download_and_parse_game}
\title{download and process an NFL game from the internet}
\usage{
download_and_parse_game(game_id, check_url = TRUE,
  return_all_data = FALSE)
}
\arguments{
\item{game_id}{the unique game identifier}

\item{check_url}{check whether the URL for the game is valid}

\item{return_all_data}{return a list containing the table as well as the preprocessing structures}
}
\value{
play by play table
}
\description{
download and process an NFL game from the internet
}
