% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_play_by_play.R
\name{download_game_play_by_play_data}
\alias{download_game_play_by_play_data}
\title{pull data from an nfl game}
\usage{
download_game_play_by_play_data(game_id_num, check_url = TRUE)
}
\arguments{
\item{game_id_num}{the unique game identifier}

\item{check_url}{check the validity of the URL created for the game id}
}
\value{
a tibble containing play by playstats for the specified nfl season
}
\description{
pull data from an nfl game
}
