% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{download_game_raw_data}
\alias{download_game_raw_data}
\title{Fetch the game file structure from the NFL website}
\usage{
download_game_raw_data(game_id, check_url = TRUE)
}
\arguments{
\item{game_id}{the game_id whose data is to be fetched}

\item{check_url}{confirm that the url can be reached}
}
\value{
the JSON data from the NFL converted into an R readable format
}
\description{
Fetch the game file structure from the NFL website
}
