% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_play_by_play.R
\name{download_week_play_by_play_data}
\alias{download_week_play_by_play_data}
\title{pull data from an nfl week}
\usage{
download_week_play_by_play_data(season, type = "reg", week = NULL)
}
\arguments{
\item{season}{the year for which you want stats}

\item{type}{pre or reg whether you want preseason or regular season stats}

\item{week}{the week of the season for wich you want stats}
}
\value{
a tibble containing play by playstats for the specified nfl season
}
\description{
pull data from an nfl week
}
