% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_play_by_play.R
\name{find_game_ids}
\alias{find_game_ids}
\title{retrieve the individual game ID's from the NFL schedules}
\usage{
find_game_ids(season, type, week = NULL, check_url = TRUE)
}
\arguments{
\item{season}{the year}

\item{type}{pre reg or post}

\item{week}{the week of the season}
}
\value{
an integer vector containing the game id's from the NFL schedule page
}
\description{
retrieve the individual game ID's from the NFL schedules
}
