% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mappings.R
\name{find_stats_with_name}
\alias{find_stats_with_name}
\title{return all stats that contain a certain term in their name}
\usage{
find_stats_with_name(part_of_name)
}
\arguments{
\item{part_of_name}{string to search stats for}
}
\value{
named integer vector with stats that contain part_of_name
}
\description{
return all stats that contain a certain term in their name
}
