% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_environment.R
\name{generate_game_environment}
\alias{generate_game_environment}
\title{downloads the raw game data and creates preprocessing structures}
\usage{
generate_game_environment(game_id, check_url = TRUE)
}
\arguments{
\item{game_id}{the game id for which the environment is to be created}

\item{check_url}{whether or not to check the validity of the url created from the game id}
}
\value{
an environment that contains data structures to be used when processing the game
}
\description{
downloads the raw game data and creates preprocessing structures
}
