% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_environment.R
\name{generate_game_environment_from_raw_data}
\alias{generate_game_environment_from_raw_data}
\title{create preprocessing structures from raw game data}
\usage{
generate_game_environment_from_raw_data(raw_game_data, game_id = NULL)
}
\arguments{
\item{raw_game_data}{the list structure returned from download_game_raw_data}

\item{game_id}{the id number for the game.}
}
\value{
a group of data structures that will be used when creating the game play by play
}
\description{
create preprocessing structures from raw game data
}
