% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{generate_game_play_by_play_data}
\alias{generate_game_play_by_play_data}
\title{Extract variables from every play and merge them into a data frame}
\usage{
generate_game_play_by_play_data(game_environment)
}
\arguments{
\item{game_environment}{}
}
\value{
data frame containing the variable that are processed from every play
}
\description{
Extract variables from every play and merge them into a data frame
}
\keyword{internal}
