% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_data.R
\name{generate_play_data}
\alias{generate_play_data}
\title{parse the data for a given play}
\usage{
generate_play_data(play_number, play, game_environment)
}
\arguments{
\item{play_number}{the numeric index of the play}

\item{play}{the play for which to generate stats}

\item{game_environment}{the game}
}
\value{
a row of the game play by play dataframe representing the play.
it includes indicators for whether  or not certain events occurred
}
\description{
parse the data for a given play
}
\keyword{internal}
