% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_environment.R
\name{generate_play_stat_matrix_list}
\alias{generate_play_stat_matrix_list}
\title{Create a list of compact matrices to store the player stats for each play}
\usage{
generate_play_stat_matrix_list(play_list, play_count = length(play_list),
  player_id_hash_table)
}
\arguments{
\item{play_list}{list of plays in the game pulled from the NFL's JSON}

\item{play_count}{the number of plays in the game}

\item{player_id_hash_table}{hasmap that maps the NFL player_id's to rows in the player_matrix}
}
\value{
a list containing matrices of the player stats for each play in the game.
each matrix has 3 columns:
player_matrix_row_num - the row in the player_matrix which corresponds to the player who earned this stat
stat_id - The integer code that the NFL uses to designate different stat events
yards - the yards associated to this stat
the rows of the matrix are in order of the sequence specified in the NFL JSON
}
\description{
Create a list of compact matrices to store the player stats for each play
}
\keyword{internal}
