% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_environment.R
\name{generate_player_table}
\alias{generate_player_table}
\title{create a compact matrix to hold all player data needed to process the game}
\usage{
generate_player_table(home_team, away_team, play_list)
}
\arguments{
\item{home_team}{the abbreviation of the home team}

\item{away_team}{the abbreviation of the away team}

\item{play_list}{the list of plays for the game}
}
\value{
a list containing
(1) player_matrix
contains the player id, name, and team for each unique player who recorded a stat in the game.
(2) player_id_hash_table
maps the NFL player id string to the appropriate row of the player_matrix.
(3) num_players
the number of players in the player_matrix
}
\description{
create a compact matrix to hold all player data needed to process the game
}
\keyword{internal}
