% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{get_json_url}
\alias{get_json_url}
\title{Create the url to fetch the game json data}
\usage{
get_json_url(game_id, check_url = TRUE)
}
\arguments{
\item{game_id}{the game_id for which the url will be created}

\item{check_url}{confirm that the url can be reached}
}
\value{
the url for the game JSON
}
\description{
Create the url to fetch the game json data
}
