% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_utilities.R
\name{initialize_empty_play_data}
\alias{initialize_empty_play_data}
\title{Create the single row data frame for a play with all variables}
\usage{
initialize_empty_play_data()
}
\value{
single row data frame with variables set to 0 or NA as appropriate.
}
\description{
Create the single row data frame for a play with all variables
}
\keyword{internal}
