% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_utilities.R
\name{initialize_NA_play_data}
\alias{initialize_NA_play_data}
\title{Create the single row data frame for non plays}
\usage{
initialize_NA_play_data()
}
\value{
single row data frame with NA and 0 filled out for non plays
}
\description{
Create the single row data frame for non plays
}
\keyword{internal}
