% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{parse_raw_game_data}
\alias{parse_raw_game_data}
\title{convert raw game data into table}
\usage{
parse_raw_game_data(raw_data, return_all_data = FALSE)
}
\arguments{
\item{raw_data}{raw game data list created by RJSONIO from the NFL JSON}

\item{return_all_data}{return auxililary data structures in addition to the table}
}
\value{
play by play table
}
\description{
convert raw game data into table
}
