% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_play_by_play_modifiers.R
\name{quarter_end_status}
\alias{quarter_end_status}
\title{detect wheter or not the game is at the end of a quarter}
\usage{
quarter_end_status(game_description)
}
\arguments{
\item{game_description}{description of the play from play by play data}
}
\value{
TRUE or FALSE if the game is at the end of a quarter
}
\description{
detect wheter or not the game is at the end of a quarter
}
\keyword{internal}
