% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_game.R
\name{select_and_order_for_nflscrapR_ep_wp}
\alias{select_and_order_for_nflscrapR_ep_wp}
\title{Put the columns in the order necessary for the ep and wp functions}
\usage{
select_and_order_for_nflscrapR_ep_wp(game_pbp)
}
\arguments{
\item{game_pbp}{the play by play table}
}
\value{
the play by play table with the columns in the correct positions for the ep and wp functions
}
\description{
Put the columns in the order necessary for the ep and wp functions
}
\keyword{internal}
