% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_play_by_play_modifiers.R
\name{select_columns_nflscrapR_final_order}
\alias{select_columns_nflscrapR_final_order}
\title{arrange the play by play table to match output from nflscrapR}
\usage{
select_columns_nflscrapR_final_order(game_play_by_play)
}
\arguments{
\item{game_play_by_play}{play by play table}
}
\value{
the play by play table with only the columns returned by nflscrapR in the order returned by nflscrapR
}
\description{
arrange the play by play table to match output from nflscrapR
}
\keyword{internal}
