% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_data.R
\name{stat_in}
\alias{stat_in}
\title{match integer stat id's to named stats}
\usage{
stat_in(play_stat_id_vector, stats_to_check_for_vector)
}
\arguments{
\item{play_stat_id_vector}{integer vector containing NFL stat id's of the stats accrued during a play}

\item{stats_to_check_for_vector}{character vector containing names of stats to check for}
}
\value{
a logical vector that tells which elements of the stat id vector are contained in the stat names
}
\description{
match integer stat id's to named stats
}
\keyword{internal}
